+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Items with description will have an tooltip 
 - Bug fix: Add compendiums to merged compendiums will trigger an scroll
 - New batch actions for GM: Mark All as On Sale, Remove All from Sale, Show All, Hide All, Make All Purchasable, Disable Purchase for All, Set All as Saleable, Remove All from Sale
+- you can create headers for items in the general tab
 
 # 0.1.1
 - Performance fix: the store loads 3,04 faster.